+++ creature.h
@@ -11,7 +11,7 @@
 using namespace std;
 
 Item* loaditem(string name);
-void text(string,int,int,WORD);
+void text(string,short,short,WORD);
 void clear();
 void num(int num,int X, int Y,WORD color);
